Drop privileges to keystone:keystone
Drop privileges to the keystone user, a member of the keystone
group, prior to executing any command defined by this snap.

If the user and group don't already exist, they will be created.
Additionally, file permissions and ownership of setup files are
adjusted to properly limit access.

This change also includes a minor path name fix for templates.

Change-Id: If9bf24fc65412b90b8b1890944a469de23888c32