Add install setup key and drop symlinks
The install setup key is introduced to determine the root directory
where setup dirs, templates, and copyfiles are installed. We set
the install key to 'classic' in snap-openstack.yaml, which results
in using / as the root directory. This is used in favor of symlinks
which are dropped in this change.

Additionally PYTHONPATH is updated as that is required in later
snapcraft versions, and /etc/keystone/keystone.conf.d is moved
to /etc/keystone/conf.d to align with the other OpenStack snaps.

Change-Id: I62ba36d6f3efd3493a8e268bf1babfe232b41764