Add configure hook and install-method config
The install-method config is introduced to detetermine the
root directory where setup dirs, templates, and copyfiles are
installed. If install-method=classic (default) the directories
are installed as specified in snap-openstack.yaml, with / as the
root directory. If install-method=strict the directories are
installed as specified in snap-openstack.yaml, with $SNAP_COMMON
as the root directory.

The install-method config can be set as follows:
  snap set keystone install-method=classic
  snap set keystone install-method=strict

Additionally PYTHONPATH is updated as that is required in later
snapcraft versions, and /etc/keystone/keystone.conf.d is moved
to /etc/keystone/conf.d.

Change-Id: I62ba36d6f3efd3493a8e268bf1babfe232b41764