Use keystone.conf from both SNAP and SNAP_COMMON
As part of this change, the pyargv uwsgi option is dynamically
created. This is needed because specifying hard-coded config
files via the uwsgi pyargv option caused failures when a file
didn't exist. Now, when dynamically creating pyargv, if an
OpenStack config file doesn't exist, it won't be added.

This change requires that the uwsgi templated configuration be
rendered as part of the entry point, rather than globally for
the snap.

Change-Id: I99541fbf1292a3a4d118c0a8da8dd34891337c88